The phpspec.php command line script is more versatile now. It uses the directory of the script itself (__DIR__) to load the UniversalClassLoader and also passes the same dir to the UniversalClassLoader::registerNamespace method as the second argument.
This allows the user to clone (or download) the source and simply
make a link to the phpspec.php file into /usr/bin/phpspec and be
done with it.

If the following were not done, the following error occurs.

PHP Warning:  require_once(PHPSpec/Loader/UniversalClassLoader.php):
failed to open stream: No such file or directory in
/home/likewise-open/ALLIEDINSURE/amjad/src/phpspec/scripts/phpspec.php on line 175

Mainly because the user won't run phpspec.php in the
./phpspec/src/ directory