Use nightly instead of 7.1 until 7.1 is official
I'm not sure how I feel about this yet, my other open PR is failing
because RC2 is failing the integration tests, that is also the cause
of the failing test on #1000 which @jameshalsall is working on.

I think it makes sense to test against the nightly builds since it
reveals issues earlier (for example, the RFC for disallowing null
parameters into `get_class` that passed 15/3 on October 9th has
already revealed some changes that are required in PhpSpec, I
have configured Travis to not panic too much about failures in
nightly builds because it is also likely to cause some false
positives.) – I think official minor builds should be added once
officially released.

If this is merged, the other two open PRs should start passing.