Update install instructions
The current instructions detail how to install Composer which is already explained right before it. It seems this is just an oversight as this line should explain how PhpSpec itself is installed through Composer. I also updated the Composer binary to the global one as that's the default when installing Composer.